Perbaikan untuk tipe data fungsi pembagian

+++ fungsiSwitch2/fungsiSwitch2.cpp
@@ -18,7 +18,7 @@ int perkalian(int a, int b)
     return a * b;
 }
 
-int pembagian(int a, int b)
+float pembagian(float a, float b)
 {
     return a / b;
 }
